Add product branch

+++ pages/api/products/[id].js
@@ -2,7 +2,8 @@ import dbConnect from "../../../utils/mongo.js";
 import Product from "../../../models/Product.js";
 
 export default async function handler(req, res) {
-    const { method, query: { id } } = req;
+    const { method, query: { id }, cookies } = req;
+    const token = cookies.token
 
 
     dbConnect();
@@ -16,6 +17,9 @@ export default async function handler(req, res) {
         }
     }
     if (method === "PUT") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not autentificated!");
+        }
         try {
             const product = await Product.findByIdAndUpdate(id, req.body, {
                 new: true,
@@ -27,6 +31,9 @@ export default async function handler(req, res) {
     }
 
     if (method === "DELETE") {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json("Not autentificated!");
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(200).json("The product has been deleted!");
